pyproject.toml
[tool.poetry]
name = "task-1"
version = "0.1.0"
description = ""
authors = ["Ahmed El-Zahaby <ahmedelzahaby8122000@gmail.com>"]
readme = "README.md"
packages = [{include = "task_1"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
pytz = "^2022.6"
image-kit = "^0.0.3"
djangorestframework = "^3.14.0"
pillow = "^9.3.0"
django-imagekit = "^4.1.0"
django-rest-knox = "^4.2.0"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
